chore: move changelog to deprecated tracker docs
When the deprecated tracker endpoints are removed we should only remove
a couple of lines from the new tracker docs. The new tracker docs should
be written as if old tracker did not exist. Having a changelog style at
the start of the new tracker docs makes them hard to read. Devs not
having the knowledge of the old endpoints should not have to acquire it.

Move the changelog into a migration section in the deprecated tracker
docs. Add a similar disclaimer on both pages just to make sure readers
land on the right page now that we switched their order. When old
tracker is removed we only need to remove the disclaimer in the
tracker.md.