docs: stick to one word for validate
We should stick to using one word for a certain behavior. If the word we
use is not describing the behavior well we should change it. Only if we
cannot due to backwards compatibility should we use another word to
improve the description.

Dry-run might imply that we are able to report the objects that will be
created/updated/deleted which we do not.